+++ src/Types/UserIdentificationHintType.php
@@ -1,0 +1,71 @@
+<?php
+//
+// Copyright (C) 2020 Authlete, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+// either express or implied. See the License for the specific
+// language governing permissions and limitations under the
+// License.
+//
+
+
+/**
+ * File containing the definition of UserCodeCharset class.
+ */
+
+
+namespace Authlete\Types;
+
+
+use Authlete\Util\LanguageUtility;
+
+
+/**
+ * Types of hints for end-user identification.
+ *
+ * @since 1.8
+ */
+class UserIdentificationHintType
+{
+    use EnumTrait;
+
+
+    /**
+     * An ID token previously issued to the client.
+     *
+     * @static
+     * @var UserIdentificationHintType
+     */
+    public static $ID_TOKEN_HINT;
+
+
+    /**
+     * An arbitrary string whose interpretation varies depending on contexts.
+     *
+     * @static
+     * @var UserIdentificationHintType
+     */
+    public static $LOGIN_HINT;
+
+
+    /**
+     * A token whose format is deployment or profile specific.
+     *
+     * @static
+     * @var UserIdentificationHintType
+     */
+    public static $LOGIN_HINT_TOKEN;
+}
+
+
+// Call UserIdentificationHintType::initialize().
+LanguageUtility::initializeClass(__NAMESPACE__ . '\UserIdentificationHintType');
+?>